Moved bundle management to admin section
