+++ app/views/admin/show.html.erb
@@ -6,7 +6,7 @@
 
 <div data-no-turbolink class="settings-tabs">
   <ul data-no-turbolink>
-    <% ['application_settings', 'user_management', 'application_status', 'bundle'].each do |target_id| %>
+    <% ['application_settings', 'user_management', 'bundles', 'application_status'].each do |target_id| %>
       <li><a href = '#<%= target_id %>'><%= target_id.humanize.titleize %></a></li>
     <% end %>
   </ul>
@@ -48,11 +48,12 @@
     <%= render partial: "user_list" %>
   </div>
 
+  <div id="bundles" >
+    <%= render partial: "bundle_list"%>
+  </div>
+
   <div id="application_status" >
     <h2>This feature is coming soon</h2>
   </div>
 
-  <div id="bundle" >
-    <h2>This feature is coming soon</h2>
-  </div>
 </div>
