--- app/views/bundles/edit.html.erb
@@ -1,26 +1,0 @@
-<div class="panel panel-default">
-<div class="panel-heading"><h1>Edit Bundle : <%= @bundle.title %></h1></div>
-
-<%= bootstrap_form_for @bundle, url: bundle_path(@bundle), html: {id: "edit_bundle_form"} do |f| %>
-  <div class="panel-body">
-    <%= f.text_field :title, label: 'Bundle Title' %>
-
-
-    <%= f.form_group :active do %>
-      <%= f.check_box :active, label: 'Default' %>
-    <% end %>
-
-  </div>
-
-  <div class="panel-footer">
-    <%= f.submit 'Update Bundle', :class => "btn btn-success", :id => "submit_button" %>
-    <%= submit_tag "Cancel", :class => "btn btn-default", :type => "button", :onclick => "history.back()" %>
-  </div>
-<% end %>
-
-</div>
-
-
-<%= render partial: 'remove_modal' %>
-
-<%= render partial: 'remove_panel', locals: { name: @bundle.title, type: 'bundle', message: 'Removing a bundle will also delete all associated products, product tests, and test execution results. Be sure you want to do this.', delete_path: bundle_path(@bundle) } %>
