+++ config/cypress.yml
@@ -45,8 +45,6 @@ auto_approve: true
 auto_associate_pocs: false
 
 auto_confirm: true
-# setting to disable the bundles page until admin roles are available
-disable_bundle_page: false
 
 file_upload_root: "data/upload/"
